% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot_country}
\alias{barplot_country}
\title{Events barplot by country}
\usage{
barplot_country(data, countries, type = "events", min_year = NULL,
  max_year = NULL, na_omit = TRUE)
}
\description{
Events barplot by country
}
