% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_map}
\alias{make_map}
\title{Make a map}
\usage{
make_map(data, type = "events", area = NULL, color = "red",
  scale = NULL, line_color = "white", line_width = 0.8,
  faceted = FALSE, world = NULL)
}
\description{
Make a map
}
