% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{makeMap}
\alias{makeMap}
\title{Make a map}
\usage{
makeMap(data, type = "events", area = NULL, color = "red",
  scale = NULL, lineColor = "white", lineWidth = 0.8,
  faceted = FALSE)
}
\description{
Make a map
}
